Removed is_writable log file checking
If the log file is not writable, PHP will throw an error anyway: this code is
unnecessary. Furthermore, simply checking to see if the file is writable is
inadequate: the file may not yet exist, but may be in a writable directory.
Writing to the file would succeed, but the is_writable check would fail.